Health and hunger rework, small changes.
Health is handled as a float now. Hunger drains twice as slow now and is also handled as a float that updates every frame instead of n seconds. Food spawns slightly less now because of hunger rework. Mobs no longer are healed when sent to the ether. Basicmob attacks are slower. Dodges tweaked, invincible on startup, and travel faster. Alt + F4-ing on day 1 will not save the game, useful for testing. Sealed kilns will drain fuel 3x as fast. Small bug fixing inventories inside of items being lost when picking up with a full inventory. Loot drop off by 1 error fixed. Fixed lynchers spawning in the ether. New journal winter entry. New sounds & ambience added.

+++ Assets/Scripts/DamageArgs.cs
@@ -7,7 +7,7 @@ public class DamageArgs : EventArgs
 {
     public string damageSenderTag;
     public GameObject senderObject;
-    public int damageAmount;
+    public float damageAmount;
     public DamageType dmgType;
 }
