+++ Assets/Scripts/UI/HungerBar.cs
@@ -7,13 +7,13 @@ public class HungerBar : MonoBehaviour
 {
     public Slider slider;
 
-    public void SetMaxHunger(int _hunger)
+    public void SetMaxHunger(float _hunger)
     {
         slider.maxValue = _hunger;
         slider.value = _hunger;
     }
 
-    public void SetHunger(int _hunger)
+    public void SetHunger(float _hunger)
     {
         slider.value = _hunger;
     }
